Martial arts always gets its skill bonuses added
It's the lowest-damaging out of the various options, in part because of
a stupid thing like even at grand master skill, 25% of their attacks get
no skill bonuses. So, add the bonus (and exercise skill) on every attack
that hits. I think that even with this adjustment, it will remain the
lowest damaging option.

I considered another option in which you could still fail to exercise
based on getting a crit-fail dieroll, but it occurred to me that this
would have virtually no effect until the late game when you ought to be
hitting regardless of dieroll.