Initial port of SLASHEM techniques and lycanthrope race.
This commit is extremely large, since it incorporates the entire technique system for SLASHEM. This also required porting a few SLASHEM features, namely medical kits and ghasts. The lycanthrope race in Splice was changed to be much closer to its counterpart in SLASHEM. I did not want to bundle these commits together, but a good deal of lycanthrope code deals with techniques. This commit also contains a restructuring of merfolk, because it needed to be changed at the same time as lycanthropes.