Race reflagging
This commit fixes several long-running bugs in SpliceHack relating to races, race selection, and cannibalism. Most players will not notice the difference, but this will make it much easier to add more races in the future. The race selection has been changed again slightly, hopefully for the last time. The code for this commit is based on the race refactoring in dnethack, with some minor changes to make it compatible with vanilla 3.6.2.