Added new fey.
This commit adds the red cap and the baobhan-sith, two new fey creatures. Red caps are not terribly interesting, but baobhan-sith are somewhat unique. Female players killed by them rise as baobhan-sith, but male and nonbinary players do not. Additionally, they can rarely drop the dresses they wear, which are a new item.