Ported Serpent Tongue and the Bat from Hell.
The Bat from Hell may actually be somewhat balanced in Splice. The artifact damage calculation differences between vanilla and SLASHEM mean that the damage output will be less absurd, and it has been changed so that it is not guaranteed for rogues in order to avoid unbalancing classes. The only way it can be randomly generated is in the already-rare junk shop, meaning that finding it is almost impossible.