Added Encyclopedias.
This is an archived YANI by Luxidream, FIQ, and aosdict. Encyclopedias can only be read once. They identify a number of items deterined by the BUC status of the encyclopedia. These items are selected from a list of many mangic items and items with randomized appearances. There is no guarantee that the items identified will be ones that the player does not know.