Added spiked chains.
Spiked chains, in all their overpowered glory. The code for them is very hacky at the moment, but they work as intended. Essentially they are polearms, but with none of the downsides. You can use them in melee or at a reach and they will work just as well. Finally, a reason to invest in the flail skill.