Replace Medusa level titan
The titan was infamous for making that one Medusa level variant much
harder than the other ones, mostly because it could almost always summon
nasties.

Replace it with two golden nagas (keeping the alignment balance the
same: lawful characters will have an easier time of this level since
they'll probably be peaceful), one where the titan used to be and one on
an island outside. As spellcasters, they provide the same type of
challenge, and as nagas they fit Medusa thematically.