Fixed food spontaneously combustin when the player was interrupted while eating.
This is a pretty hacky fix, and I am not terribly proud of it. It turns out that orotten was defined as oeroded as well, despite the fact that it only ever contains a boolean value and oeroded is a bitfield of size 2. This was causing issues with cooking, which is defined using oeroded as well. This fix probably will not be permanent, but it should resolve the bug in the short term.