Set 'checkpoint' option to true by default
Unrecoverable crashes were happening with the 'checkpointing not in
effect' issue. I discovered that recovery becomes possible again if
checkpoint is set to true, so this commit makes true the default. It's
still configurable by the user, if anyone is really unhappy with the
additional time it takes to save the level to disk when changing levels,
but setting it to false basically indicates "yes, I'm okay with my game
being unrecoverable if it crashes".

(cherry picked from commit f2753e9f786af5c32f31a3de899c8a1213a243f3)