Kicking a tame monster is always "clumsy"
This is intended to prevent cartomancer unique death abuse in JNH. 

It might be fun to extend this into providing a useful reason to kick pets: it causes them to wake up (if asleep) and make one random move, if they're standing in a spot and you'd prefer them to be in a different spot.  It shouldn't abuse tameness in that case.