Fix general case of ignored wish quantities
A specified quantity only is respected if it's greater than 1, but
mksobj(mkobj.c) sometimes generates certain items (e.g. kelp fronds) in
stacks of multiple items. This commit changes the specified quantity to
be explicitly respected as usual even if it's 1, so that a wish for `1
kelp frond' will not return 2 kelp fronds 50% of the time.