Add material interactions when forcing a lock
A disparity between strengths of the materials of a target box and the
weapon used for prying open a lock will now affect the chances of the
weapon breaking. Likewise, the comparitive strength of the box material
(vs the base material of the container) will have an impact on the
chances of the box itself shattering when attacked with a blunt weapon.