Make taming cats with thrown catnip easier
Putting catnip taming in uhitm.c means it's hard to actually hit a cat
with thrown catnip, which is very difficult compared to standard animal
taming (e.g. throwing a tripe ration to a dog); since there is already a
downside to using catnip -- confusing the cat that gets hit with it --
it seems needless to make it tough to throw successfully on top of that.