Allow forcing monsters to be male via level file
It was already possible to force a monster to be female by including
female=1 in their des.monster() declaration, but female=0 only meant sex
was randomized as usual.

With this change omitting the female option entirely leaves the sex
randomized, while setting female=0 forces the generated monster not to
be female.