Avoid overriding orientation set in rcfile
Because straight orientation is 0, checking for a nonzero
flags.initorientation simply overrides the hero's sexual orientation
with a random one if and only if they are straight. I don't think we
even really need to check orientation in u_init, since a random
orientation has already been selected in optfn_orientation(options.c) if
the player didn't specify one in the config file, or in
str2orientation(role.c) if the player specified `@' or `*'.

But if we do want to verify the player has gotten their orientation set
at this point, checking for a nonzero initorientation won't do it
properly.