Change blessed playing card luck message
It seems a little silly to me that the hero would take the time to
shuffle the playing card deck up to 13 times in a row. Since there are
13 different ranks of cards in a playing card deck, and the item is
themed around playing cards anyway, I changed it so that the high card
(or the kicker when the hand doesn't use all 5 cards) represents the
current value of Luck. So a player with 2 luck will get a message like:

    You draw a hand of five cards. Two pair! The kicker is the two of
    clubs.

While a player with 13 luck will get:

    You draw a hand of five cards. Wow, a royal flush! The high card is
    the king of spades.