Improve kicking to be a better fallback plan
"Martial" players (monk, samurai, or bigfoot) are immune to some of the
frustrations in the kicking system, such as injured legs or clumsiness.

Martial players' mighty feet can kick apart most dungeon features.
Not implemented: kicking thrones apart. Seems too mean.

Kicking the walls is treated as more of a "test kick", potentially
revealing information about secret doors/walls. This is instead of
assuming that the player is intending to kick a solid object with
all their might, leading only to injury and woe.