Reformat role.c
Put standard species on one line
splicehack species on the next
genders on the next
alignments on the next

add a #define for the common use case of "all 3 [genders|alignments]"

This should make it harder to confuse neutral and nonbinary flags.
The #defines will work nicely if further genders or alignments are added.