Improve character creation in window ports.
This commit introduces a number of changes: First, alter the selection menu to allow for more than two roles with the same initial character. Second, display an in-progress character sheet in the curses menu during character creation. Third, refer to race as species. This final change is subject to change based on whether a more appropriate term is found. Political correctness is a neverending process, and the least I can do is take a few small steps.