Remove random secret corridors
From UnNetHack. The reasoning for doing so there (and the same here) is
that secret corridors really don't add much of anything to the game,
except the thrill of standing at the apparent dead end of a corridor and
mashing the search key to see if it really is the dead end or not.

Secret corridors still can and do generate in special levels when
they're explicitly specified, e.g. with the mapchar H. But they won't be
generated randomly as part of corridor drawing (as in Quest filler
levels that generate a bunch of random rooms and connecting corridors).