Level parser: Can specify "dead" for MONSTER
This will create the monster, then kill it. A kludge is needed so that
it doesn't count as an actual monster death (which credits the player
with vanquishing it.)

Since all the other MONSTER attributes only affect alive monsters, the
level parser will complain if you try to specify "dead" alongside any
other attributes.

The intent of this is to use it in Orcish Town to generate piles of
items along with corpses rather than scattering gear manually.

(cherry picked from commit 4eaaaa562a6fbca4c7a7dcae1e854d595707eab8)