Monster flanking AI.
With this addition, monsters with the M2_FLANK tag will attempt to
outflank the player, moving around them in order to set up flanking
situations with allies. While I could have added this behavior for
all monsters, I think only enabling it for intelligent monsters and
pack hunters is more flavorful, and helps to distinguish monsters
from one another.

Plus, we have a bunch of free M2 tags now and we might as well use
them :)