Fixed errors exposed by -Wall.
For whatever reason, I was using the linux hint file instead of
linux-debug, and as such there were a lot of little bugs flying
under the radar. This commit cleans up most of them; the ones that
are left are pretty much all from vanilla. Thanks K2 for pointing
this out and messaging me!