Fix windows compilation.
As discovered by community member Doesnt, there is a palette issue
with the current ridingmark, templatemark, and splash, which breaks
windows compilation for 32 bit systems. As a stopgap, these files
have been removed, and replaced with their vanilla counterparts,
or, failing that, the vanilla pilemark.

In the future, we will add unique template, riding, and splash tiles
again.