Make hallucination and EotO block gazes
The Eyes of the Overworld occupy the same slot as blindfolds and towels,
preventing the player from self-blinding when up against a gaze attack.

This removes that quirky hidden weakness and replaces it with a boon:
gaze attacks are cancelled by the magical lenses.

In many variants, hallucination also block gazes, which vanilla ended up
using but with a 25% failure rate. This effectively makes it so that
players will not use it as a defense, because hallucination already
comes with plenty of downside. Buff it back up to 100% effectiveness.