Unlocking tools can break.
This stemmed from a conversation over IRC. Lock picks now have a
chance of breaking when used, but are also stackable. Credit cards
have a chance of breaking upon each use, while keys degrade slowly
over time. The purpose of this change is to encourage more careful
use of resources.