Improve template application.
Despite the appearances, this is a rather small change. Previously,
attacks from templates were applied to the end of the list of the
base monster's attacks. This commit adds some logic so that template
attacks can be written in the same format as regular monster attacks,
and will be applied forward, rather than backward.