Use matching object materials for gold/silver DSM
A flavor-based nerf to silver dragon scale mail: it is metallic and weighs more, as other silver items are. This makes it unusable for silver-haters and a significant impediment for spellcasters. Non-silver-hating non-spellcasters won't be hit much by this nerf, though those characters can just as easily wear a silver shield in the first place.

Because this nerf relies upon a literal interpretation of "silver dragon", gold armor is also now heavy and metallic, though due to its unpopularity, the impact will likely be small.